+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/)
 ### Fixed
 
 - Enable editable SRP amount field only for pending and rejected requests (#51)
+- CSS classes in templates
 
 
 ## [1.11.0] - 2022-02-02